chat : Left群列表加载成功

+++ NapCatScript.Core/JsonFormat/GetJsons/get_group_list.cs
@@ -62,5 +62,5 @@ public class GroupInfo
     /// 群号
     /// </summary>
     [JsonPropertyName("group_id")]
-    public double GroupId { get; set; }
+    public long GroupId { get; set; }
 }
